refactor(CardPaymentContext): 파일 위치 및 네이밍 수정으로 인한 코드 수정
