+++ src/components/context/PaymentsProvider.tsx
@@ -1,23 +1,23 @@
 import React, { useReducer } from 'react';
 
 import reducer from '../../store/reducer';
-import initialState from '../../store/initialState';
-import { CardPaymentDispatchContext, CardPaymentStateContext } from '../../hooks/useContextHooks';
+import cardFormState from '../../store/cardFormStore';
+import { CardPaymentDispatchContext, CardPaymentStateContext } from './CardPaymentContext';
 
 type ChildProps = {
   children: React.ReactNode;
 };
 
-const PaymentProvider: React.FC<ChildProps> = ({ children }) => {
-  const [state, dispatcher] = useReducer(reducer, initialState);
+const PaymentFormProvider: React.FC<ChildProps> = ({ children }) => {
+  const [formState, formDispatcher] = useReducer(reducer, cardFormState);
 
   return (
-    <CardPaymentStateContext.Provider value={state}>
-      <CardPaymentDispatchContext.Provider value={dispatcher}>
+    <CardPaymentStateContext.Provider value={formState}>
+      <CardPaymentDispatchContext.Provider value={formDispatcher}>
         {children}
       </CardPaymentDispatchContext.Provider>
     </CardPaymentStateContext.Provider>
   );
 };
 
-export default PaymentProvider;
+export default PaymentFormProvider;
