+++ src/components/organisms/CardList/index.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import { useCardListState } from '../../../hooks/useContextHooks';
+
+import Message from '../../atomics/Message';
+import { useCardListState } from '../../context/CardPaymentContext';
+import { VStack } from '../../layout/flexbox';
 
 import CardItem from '../../molecules/CardItem';
 
@@ -10,11 +13,25 @@ const CardList: React.FC = () => {
   return (
     <CardPreviewListWrapper>
       {cardList.map((cardItem) => (
-        <CardItem key={cardItem.id} card={cardItem} />
+        <CardItemWrapper key={cardItem.id}>
+          <CardItem key={cardItem.id} card={cardItem} />
+          <Message fontWeight={700} fontSize="14px" lineHeight="16px" color="#575757" opacity={0.9}>
+            {cardItem.nickName}
+          </Message>
+        </CardItemWrapper>
       ))}
     </CardPreviewListWrapper>
   );
 };
+
+const CardItemWrapper = styled.div`
+  ${VStack}
+  align-items: center;
+
+  div + span {
+    margin-top: 12px;
+  }
+`;
 
 const CardPreviewListWrapper = styled.div`
   display: flex;
