+++ src/pages/AddCard/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import { useCardPaymentState } from '../../hooks/useContextHooks';
+import { useCardPaymentState } from '../../components/context/CardPaymentContext';
+
 import CardItem from '../../components/molecules/CardItem';
 import Header from '../../components/molecules/Header';
 import CardForm from '../../components/organisms/CardForm';
